Implemented .MinimumSwitchingPlayers property.
Added property to config.properties: Server.MinimumSwitchingPlayers. This property can be set to a numeric value >= 0. It will describe how many (recognized) players have to play on a server (and on a listening-channel) to enable switching.
Recognized means listed in users.properties.
This property is default 3 (and therefore if not set also 3).